Remove the 5-year anniversary banner from the home page (#1202)
* Remove Bitcoin Design is 5 banner from home

The reverse of #1190, removing the banner again to show the rotating banners designers in the community have come up with.

* Added link to projects page

We still should link to the report from somewhere. I added a link to the projects page, since the report is very thorough in that regard. There might also be other good places to link. If you can think of some, let me know.